Added mount for FrSky M9 gimbal to the side of the Dactyl Manuform.
This is a joystick from an R/C controller, typically used to control cars, planes, drones etc. I'm pretty
sure it can be connected to an analogue input of an Arduino and used in QMK as a joystick or hopefully
a mouse.

All of this is currently theoretical. I have measured up the gimbal, and it ought to fit, but I've not
printed any of this yet. Nor have I tested the gimbal with QMK.

Note that the M9 gimbal is a 3.3v device, so don't connect it directly to the 5v output of the
controller!

I've also put in my preferred Dactyl Manuform config, as this is the only configuration I've tested
with.